Fix a bug in HTTP/2 messages serialization
Since years we introduced HTTP/2 support, internally we
set HTTP::Message's http-version to its appropriate version.

While this allows us to do some internal arrangements for the
1.1 vs 2 distinction (e.g. handling push-promises method nicely),
unfortunately, this version was actually used to serialize
a Request/a Response, which resulted in bogus serialization we
were able to detect using the reverse proxy tests.

To avoid removing even tiny benefits of this internal distinction,
just teach the messages to serialize themselves properly always
falling back to 1.1 if it's set to 2.

This commit also returns a removed test which was claimed to be bogus before
due to this misconception.